Relocate Filter-classes
