+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedRecordTests.java
@@ -25,6 +25,8 @@ import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceOne;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceTwo;
 import io.github.chrimle.o2jrm.test.meta.configs.ConfigOption;
 import io.github.chrimle.o2jrm.test.meta.configs.Library;
+import io.github.chrimle.o2jrm.test.meta.filters.AssumptionFilter;
+import io.github.chrimle.o2jrm.test.meta.filters.Condition;
 import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
 import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
 import io.github.chrimle.o2jrm.test.utils.AssertionUtils;
