+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedSourceProvider.java
@@ -16,6 +16,7 @@
 
 package io.github.chrimle.o2jrm.tests;
 
+import io.github.chrimle.o2jrm.test.meta.filters.AssumptionFilter;
 import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
 import java.lang.reflect.Method;
 import java.util.Arrays;
